+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/Shadow/ScreenSpaceShadowManagerDirectional.cs
@@ -64,7 +64,6 @@ namespace UnityEngine.Rendering.HighDefinition
             // Input Buffers
             public RTHandle depthStencilBuffer;
             public RTHandle normalBuffer;
-            public ComputeBuffer lightData;
 
             // Intermediate buffers
             public RTHandle directionBuffer;
@@ -85,7 +84,6 @@ namespace UnityEngine.Rendering.HighDefinition
             // Input Buffers
             rtsdtResources.depthStencilBuffer = m_SharedRTManager.GetDepthStencilBuffer();
             rtsdtResources.normalBuffer = m_SharedRTManager.GetNormalBuffer();
-            rtsdtResources.lightData = m_LightLoopLightData.directionalLightData;
 
             // Intermediate buffers
             rtsdtResources.directionBuffer = directionBuffer;
@@ -139,9 +137,6 @@ namespace UnityEngine.Rendering.HighDefinition
                 rtsdtParams.shaderVariablesRayTracingCB._RaytracingNumSamples = rtsdtParams.numShadowSamples;
                 ConstantBuffer.PushGlobal(cmd, rtsdtParams.shaderVariablesRayTracingCB, HDShaderIDs._ShaderVariablesRaytracing);
 
-                // Bind the light & sampling data
-                cmd.SetComputeBufferParam(rtsdtParams.screenSpaceShadowCS, rtsdtParams.directionalShadowSample, HDShaderIDs._DirectionalLightDatas, rtsdtResources.lightData);
-
                 // Input Buffer
                 cmd.SetComputeTextureParam(rtsdtParams.screenSpaceShadowCS, rtsdtParams.directionalShadowSample, HDShaderIDs._DepthTexture, rtsdtResources.depthStencilBuffer);
                 cmd.SetComputeTextureParam(rtsdtParams.screenSpaceShadowCS, rtsdtParams.directionalShadowSample, HDShaderIDs._NormalBufferTexture, rtsdtResources.normalBuffer);
@@ -184,11 +179,11 @@ namespace UnityEngine.Rendering.HighDefinition
                 || !hdCamera.ValidShadowHistory(additionalLightData, dirShadowIndex, GPULightType.Directional))
                 historyValidity = 0.0f;
 
-    #if UNITY_HDRP_DXR_TESTS_DEFINE
+#if UNITY_HDRP_DXR_TESTS_DEFINE
             if (Application.isPlaying)
                 historyValidity = 0.0f;
             else
-    #endif
+#endif
                 // We need to check if something invalidated the history buffers
                 historyValidity *= EvaluateHistoryValidity(hdCamera);
 
@@ -204,12 +199,9 @@ namespace UnityEngine.Rendering.HighDefinition
             float historyValidity = EvaluateHistoryValidityDirectionalShadow(hdCamera, dirShadowIndex, m_CurrentSunLightAdditionalLightData);
 
             // Grab the history buffers for shadows
-            RTHandle shadowHistoryArray = hdCamera.GetCurrentFrameRT((int)HDCameraFrameHistoryType.RaytracedShadowHistory)
-                ?? hdCamera.AllocHistoryFrameRT((int)HDCameraFrameHistoryType.RaytracedShadowHistory, ShadowHistoryBufferAllocatorFunction, 1);
-            RTHandle shadowHistoryValidityArray = hdCamera.GetCurrentFrameRT((int)HDCameraFrameHistoryType.RaytracedShadowHistoryValidity)
-                ?? hdCamera.AllocHistoryFrameRT((int)HDCameraFrameHistoryType.RaytracedShadowHistoryValidity, ShadowHistoryValidityBufferAllocatorFunction, 1);
-            RTHandle shadowHistoryDistanceArray = hdCamera.GetCurrentFrameRT((int)HDCameraFrameHistoryType.RaytracedShadowDistanceValidity)
-                ?? hdCamera.AllocHistoryFrameRT((int)HDCameraFrameHistoryType.RaytracedShadowDistanceValidity, ShadowHistoryDistanceBufferAllocatorFunction, 1);
+            RTHandle shadowHistoryArray = RequestShadowHistoryBuffer(hdCamera);
+            RTHandle shadowHistoryValidityArray = RequestShadowHistoryValidityBuffer(hdCamera);
+            RTHandle shadowHistoryDistanceArray = RequestShadowHistoryDistanceBuffer(hdCamera);
 
             // Grab the slot of the directional light (given that it may be a color shadow, we need to use the mask to get the actual slot index)
             GetShadowChannelMask(dirShadowIndex, m_CurrentSunLightAdditionalLightData.colorShadow ? ScreenSpaceShadowType.Color : ScreenSpaceShadowType.GrayScale, ref m_ShadowChannelMask0);
@@ -271,11 +263,12 @@ namespace UnityEngine.Rendering.HighDefinition
             return sssdParams;
         }
 
-        static void ExecuteSSShadowDirectional(CommandBuffer cmd, SSShadowDirectionalParameters sssdParams, RTHandle textureArray)
+        static void ExecuteSSShadowDirectional(CommandBuffer cmd, SSShadowDirectionalParameters sssdParams, RTHandle normalBuffer, RTHandle textureArray)
         {
             // If it is screen space but not ray traced, then we can rely on the shadow map
             // WARNING: This pattern only works because we can only have one directional and the directional shadow is evaluated first.
             CoreUtils.SetRenderTarget(cmd, textureArray, depthSlice: sssdParams.depthSlice);
+            cmd.SetGlobalTexture(HDShaderIDs._NormalBufferTexture, normalBuffer);
             HDUtils.DrawFullScreen(cmd, s_ScreenSpaceShadowsMat, textureArray);
         }
 
@@ -296,7 +289,7 @@ namespace UnityEngine.Rendering.HighDefinition
                     else
                     {
                         SSShadowDirectionalParameters sssdParams = PrepareSSShadowDirectionalParameters();
-                        ExecuteSSShadowDirectional(cmd, sssdParams, m_ScreenSpaceShadowTextureArray);
+                        ExecuteSSShadowDirectional(cmd, sssdParams, m_SharedRTManager.GetNormalBuffer(), m_ScreenSpaceShadowTextureArray);
                     }
                 }
             }