+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDProfileId.cs
@@ -16,6 +16,7 @@ namespace UnityEngine.Rendering.HighDefinition
         DenoiseSSAO,
         UpSampleSSAO,
         ScreenSpaceShadows,
+        ScreenSpaceShadowsDebug,
         BuildLightList,
         ContactShadows,
         BlitToFinalRTDevBuildOnly,
@@ -118,6 +119,8 @@ namespace UnityEngine.Rendering.HighDefinition
         // RT Shadows
         RaytracingDirectionalLightShadow,
         RaytracingLightShadow,
+        RaytracingSphereLightShadow,
+        RaytracingAreaLightShadow,
         // RTGI
         RaytracingIndirectDiffuseDirectionGeneration,
         RaytracingIndirectDiffuseEvaluation,
@@ -127,6 +130,8 @@ namespace UnityEngine.Rendering.HighDefinition
         RaytracingSSS,
         RaytracingSSSTrace,
         RaytracingSSSCompose,
+        // RTShadow
+        RaytracingWriteShadow,
         // Other ray tracing
         RaytracingDebugOverlay,
         RayTracingRecursiveRendering,