+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -181,6 +181,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added render graph support of RTAO and required denoisers.
 - Added render graph support of RTGI.
 - Supporting RTSSS and Recursive Rendering in the render graph mode.
+- Supporting directional RT and screen space shadow for render graph.
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)